+++ src/extract/mod.rs
@@ -13,6 +13,7 @@ pub mod ace;
 pub mod adf;
 pub mod alz;
 pub mod arc;
+pub mod arc_conv;
 pub mod aspack;
 pub mod bcm;
 pub mod bitrock;